Maya Animation
Trying to fix some things on my animation that I could have done better. The left wrist keeps bopping from 12 to 13 when I literally have no frame on 12 for that wrist. It should just smooth out from frame 7 to 13 but it waits until frame 12 to change for whatever reason.  Will come back to this to try and fix it.