Maya Edits
My animation work for some reason is not the same work I had before (the arm is way off, going in weird directions) and it's really frustrating. I think I'm going to keep my final animation and this one separate to see if that helps because I don't know what else to do.